Add code to escape non-escaped spaces in options
Not sure if this belongs in grunt-rsync or rsyncwrapper, but here’s the idea:

Glob patterns aside, given that grunt-rsync (and rsyncwrapper) are attempts to expose the utility of rsync in a Node environment, I believe there’s a case to me made for automatically ensuring that “Node-like path strings” are correctly handled by these tools.

This is not currently the case with regards to space characters in file paths (passed as, say, the `src` argument to rsyncwrapper). Today, a JavaScript string like “/Volumes/My Disk/Source Folder” is passed directly to rsync, unquoted, without escaping space characters. This leads to rsync errors (e.g. rsync error 12, since the source path is not escaped).

Of course, it’s possible to escape space characters in a number of locations before those string reach grunt-rsync or rsyncwrapper, but when those paths may be used in other Node code that doesn’t require space escapes, those paths being passed to grunt-rsync start requiring postfixes.

Hence, this commit makes an attempt to escape space characters in any path passed to rsyncwrapper’s arguments, before handing them off to rsyncwrapper.

Meanwhile, maybe this change should be made to rsyncwrapper instead? Would be happy to discuss.